extracts new error classes to their own file
