+++ lib/measured/unit_error.rb
@@ -1,0 +1,3 @@
+module Measured
+  class UnitError < StandardError ; end
+end
